requirements.txt
joblib==1.4.2
joblib==0.17.0
matplotlib==3.5.1
numba==0.55.1
numpy==1.21.5
osgeo==0.0.1
pandas==1.3.5
plotly==5.4.0
pymoo==0.6.1.1
pyproj==3.3.0
python_docx==1.1.2
PyYAML==5.4.1
PyYAML==6.0.2
scikit_learn==1.5.2
scipy==1.8.0
Shapely==2.0.6
